requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
blis==1.2.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.11
Cython==0.29.37
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
diffusers==0.27.0
dill==0.3.8
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
et-xmlfile==1.1.0
executing==2.0.1
fastai==2.7.19
fastcore==1.7.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
geoopt @ git+https://github.com/geoopt/geoopt.git@eaadc68fcae361778edf078b503ed79e4497c071
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.65.4
hdbscan==0.8.37
higher==0.2.1
huggingface-hub==0.27.1
idna==3.7
importlib-metadata==6.11.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
openpyxl==3.1.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft @ git+https://github.com/huggingface/peft.git@9c25d9411a3db1ce9599cc5270cd07ec81ccd4c9
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ranger==0.10
ranger-fm==1.9.3
regex==2024.5.15
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
setuptools==69.5.1
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.3.4
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.19.1
torch==2.4.1
torchinfo==1.8.0
torchprofile==0.0.4
torchvision==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
triton==3.0.0
trl==0.8.6
typer==0.15.2
typing_extensions==4.13.0
tyro==0.8.4
tzdata==2024.1
umap==0.1.1
umap-learn==0.5.7
urllib3==2.2.1
wandb==0.17.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.17.2
xformers==0.0.28.post1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2
